add tooltips and address copy button
